Change mentor list view
