+++ src/index.tsx
@@ -12,7 +12,7 @@ import Dashboard from './scenes/Dashboard';
 import { Profile } from './interfaces';
 import MentorApplication from './scenes/Home/scenes/MentorApplication';
 import EditMentorApplication from './scenes/Home/scenes/EditMentorApplication';
-import RequestMentors from './scenes/Home/scenes/RequestMentors';
+import MentorList from './scenes/Home/scenes/MentorList';
 import MentorConfirmation from './scenes/Home/scenes/MentorConfirmation';
 import axios, { AxiosResponse } from 'axios';
 import { notification } from 'antd';
@@ -64,7 +64,7 @@ function App() {
             path="/program/:programId/mentor/confirmation"
             component={MentorConfirmation}
           />
-          <Route path="/program/:programId" component={RequestMentors} />
+          <Route path="/program/:programId" component={MentorList} />
         </Switch>
       </Router>
     </UserContext.Provider>
