--- src/scenes/Home/scenes/RequestMentors/components/MentorCard/interfaces.ts
@@ -1,5 +1,0 @@
-import { Mentor } from '../../../../../../interfaces';
-
-export interface MentorCardProps {
-  item: Mentor;
-}
