--- src/scenes/Home/scenes/RequestMentors/scenes/MenteeApplication/styles.css
@@ -1,7 +1,0 @@
-.textPadding {
-    margin-top: 1rem;
-}
-
-.contentMargin {
-    margin-left: 3rem;
-}
